Moves all helper functions to abstract class business
