+++ src/main/java/javaFundaments/inhertiance/Business.java
@@ -16,4 +16,36 @@ public abstract class Business {
         this.averagePrice = averagePrice;
         reviews = new ArrayList();
     }
+
+    public List getReviews() {
+        return this.reviews;
+    }
+
+    private String getPrice() {
+        String dollarSigns = "";
+        for (int i = 0; i < averagePrice; i++) {
+            dollarSigns += "$";
+        }
+        return dollarSigns;
+    }
+
+    private int getTotalStars() {
+        int totalStars = 0;
+        for (Review r : reviews) {
+            totalStars += r.getRating();
+        }
+        return totalStars;
+    }
+
+    private void updateRating() {
+        int totalStars = getTotalStars();
+        this.rating = (float) totalStars / this.reviews.size();
+    }
+
+    public Review addReview(String author, String critique, int rating) {
+        Review newReview = new Review(this, author, critique, rating);
+        this.reviews.add(newReview);
+        this.updateRating();
+        return newReview;
+    }
 }
